Fix some test, and make SchedulingResources take double parameter.

+++ src/ray/common/task/scheduling_resources.h
@@ -264,7 +264,7 @@ class SchedulingResources {
   /// \param resource_name: Name of the resource to be modified
   /// \param capacity: New capacity of the resource.
   /// \return Void.
-  void UpdateResourceCapacity(const std::string &resource_name, int64_t capacity);
+  void UpdateResourceCapacity(const std::string &resource_name, double capacity);
 
   /// \brief Delete resource from total, available and load resources.
   ///
